+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 #### [Unreleased](https://github.com/openfga/cli/compare/v0.7.0...HEAD)
 
+Added:
+- Support running `fga model test` with multiple files using glob patterns (#423)
+
 #### [0.7.0](https://github.com/openfga/cli/compare/v0.6.6...v0.7.0) (2025-06-10)
 
 > [!NOTE]